refactoring 3
